Tối ưu layout, chỉnh sửa update chức năng đăng ký, đăng nhập.
Hoàn thiện phần thống kê
