+++ resources/views/admin/dashboard/index.blade.php
@@ -9,15 +9,12 @@
                     <i class="material-icons opacity-10">weekend</i>
                 </div>
                 <div class="text-end pt-1">
-                    <p class="text-sm mb-0 text-capitalize">Today's Money</p>
-                    <h4 class="mb-0">$53k</h4>
+                    <p class="text-sm mb-0 text-capitalize">User</p>
+                    <h4 class="mb-0">{{ $userCount }}</h4>
                 </div>
             </div>
             <hr class="dark horizontal my-0">
-            <div class="card-footer p-3">
-                <p class="mb-0"><span class="text-success text-sm font-weight-bolder">+55% </span>than
-                    last week</p>
-            </div>
+
         </div>
     </div>
     <div class="col-xl-3 col-sm-6 mb-xl-0 mb-4">
@@ -28,15 +25,12 @@
                     <i class="material-icons opacity-10">person</i>
                 </div>
                 <div class="text-end pt-1">
-                    <p class="text-sm mb-0 text-capitalize">Today's Users</p>
-                    <h4 class="mb-0">2,300</h4>
+                    <p class="text-sm mb-0 text-capitalize">Product</p>
+                    <h4 class="mb-0">>{{ $productCount }}</h4>
                 </div>
             </div>
             <hr class="dark horizontal my-0">
-            <div class="card-footer p-3">
-                <p class="mb-0"><span class="text-success text-sm font-weight-bolder">+3% </span>than
-                    last month</p>
-            </div>
+
         </div>
     </div>
     <div class="col-xl-3 col-sm-6 mb-xl-0 mb-4">
@@ -47,15 +41,12 @@
                     <i class="material-icons opacity-10">person</i>
                 </div>
                 <div class="text-end pt-1">
-                    <p class="text-sm mb-0 text-capitalize">New Clients</p>
-                    <h4 class="mb-0">3,462</h4>
+                    <p class="text-sm mb-0 text-capitalize">Category</p>
+                    <h4 class="mb-0">{{ $categoryCount }}</h4>
                 </div>
             </div>
             <hr class="dark horizontal my-0">
-            <div class="card-footer p-3">
-                <p class="mb-0"><span class="text-danger text-sm font-weight-bolder">-2%</span> than
-                    yesterday</p>
-            </div>
+
         </div>
     </div>
     <div class="col-xl-3 col-sm-6">
@@ -66,15 +57,45 @@
                     <i class="material-icons opacity-10">weekend</i>
                 </div>
                 <div class="text-end pt-1">
-                    <p class="text-sm mb-0 text-capitalize">Sales</p>
-                    <h4 class="mb-0">$103,430</h4>
+                    <p class="text-sm mb-0 text-capitalize">Order</p>
+                    <h4 class="mb-0">{{ $orderCount }}</h4>
                 </div>
             </div>
             <hr class="dark horizontal my-0">
-            <div class="card-footer p-3">
-                <p class="mb-0"><span class="text-success text-sm font-weight-bolder">+5% </span>than
-                    yesterday</p>
+
+        </div>
+    </div>
+    <div class="col-xl-3 col-sm-6 mt-4">
+        <div class="card">
+            <div class="card-header p-3 pt-2">
+                <div
+                    class="icon icon-lg icon-shape bg-gradient-info shadow-info text-center border-radius-xl mt-n4 position-absolute">
+                    <i class="material-icons opacity-10">weekend</i>
+                </div>
+                <div class="text-end pt-1">
+                    <p class="text-sm mb-0 text-capitalize">Role</p>
+                    <h4 class="mb-0">{{ $roleCount }}</h4>
+                </div>
             </div>
+            <hr class="dark horizontal my-0">
+
+        </div>
+    </div>
+
+    <div class="col-xl-3 col-sm-6 mt-4">
+        <div class="card">
+            <div class="card-header p-3 pt-2">
+                <div
+                    class="icon icon-lg icon-shape bg-gradient-info shadow-info text-center border-radius-xl mt-n4 position-absolute">
+                    <i class="material-icons opacity-10">weekend</i>
+                </div>
+                <div class="text-end pt-1">
+                    <p class="text-sm mb-0 text-capitalize">Coupon</p>
+                    <h4 class="mb-0">{{ $couponCount }}</h4>
+                </div>
+            </div>
+            <hr class="dark horizontal my-0">
+
         </div>
     </div>
 @endsection
