+++ resources/views/admin/layouts/sidebar.blade.php
@@ -4,10 +4,9 @@
     <div class="sidenav-header">
         <i class="fas fa-times p-3 cursor-pointer text-white opacity-5 position-absolute end-0 top-0 d-none d-xl-none"
             aria-hidden="true" id="iconSidenav"></i>
-        <a class="navbar-brand m-0" href=" https://demos.creative-tim.com/material-dashboard/pages/dashboard "
-            target="_blank">
-            <img src="./assets/img/logo-ct.png" class="navbar-brand-img h-100" alt="main_logo">
-            <span class="ms-1 font-weight-bold text-white">Hello: {{ auth()->user()->name }}</span>
+        <a class="navbar-brand m-0" href="/">
+            {{-- <img src="./assets/img/logo-ct.png" class="navbar-brand-img h-100" alt="main_logo"> --}}
+            <h6 class="ms-1 font-weight-bold text-white">Storelly: {{ auth()->user()->name }}</h6>
         </a>
     </div>
     <hr class="horizontal light mt-0 mb-2">
