+++ resources/views/auth/login.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app')
+@extends('client.layouts.app')
 
 @section('content')
 <div class="container">
