+++ resources/views/client/home/index.blade.php
@@ -34,93 +34,7 @@
       <!-- Featured End -->
 
 
-      <!-- Categories Start -->
-      <div class="container-fluid pt-5">
-          <div class="row px-xl-5 pb-3">
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-1.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Men's dresses</h5>
-                  </div>
-              </div>
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-2.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Women's dresses</h5>
-                  </div>
-              </div>
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-3.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Baby's dresses</h5>
-                  </div>
-              </div>
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-4.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Accerssories</h5>
-                  </div>
-              </div>
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-5.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Bags</h5>
-                  </div>
-              </div>
-              <div class="col-lg-4 col-md-6 pb-1">
-                  <div class="cat-item d-flex flex-column border mb-4" style="padding: 30px;">
-                      <p class="text-right">15 Products</p>
-                      <a href="" class="cat-img position-relative overflow-hidden mb-3">
-                          <img class="img-fluid" src="img/cat-6.jpg" alt="">
-                      </a>
-                      <h5 class="font-weight-semi-bold m-0">Shoes</h5>
-                  </div>
-              </div>
-          </div>
-      </div>
-      <!-- Categories End -->
 
-      <!-- Offer Start -->
-      <div class="container-fluid offer pt-5">
-          <div class="row px-xl-5">
-              <div class="col-md-6 pb-4">
-                  <div class="position-relative bg-secondary text-center text-md-right text-white mb-2 py-5 px-5">
-                      <img src="img/offer-1.png" alt="">
-                      <div class="position-relative" style="z-index: 1;">
-                          <h5 class="text-uppercase text-primary mb-3">20% off the all order</h5>
-                          <h1 class="mb-4 font-weight-semi-bold">Spring Collection</h1>
-                          <a href="" class="btn btn-outline-primary py-md-2 px-md-3">Shop Now</a>
-                      </div>
-                  </div>
-              </div>
-              <div class="col-md-6 pb-4">
-                  <div class="position-relative bg-secondary text-center text-md-left text-white mb-2 py-5 px-5">
-                      <img src="img/offer-2.png" alt="">
-                      <div class="position-relative" style="z-index: 1;">
-                          <h5 class="text-uppercase text-primary mb-3">20% off the all order</h5>
-                          <h1 class="mb-4 font-weight-semi-bold">Winter Collection</h1>
-                          <a href="" class="btn btn-outline-primary py-md-2 px-md-3">Shop Now</a>
-                      </div>
-                  </div>
-              </div>
-          </div>
-      </div>
-      <!-- Offer End -->
 
 
       <!-- Products Start -->
